+++ commands/__init__.py
@@ -2,3 +2,5 @@ from commands import config
 from commands import init
 from commands import pull
 from commands import push
+from commands import pullSource
+from commands import pushTarget